+++ src/config/validations.config.ts
@@ -1,0 +1,3 @@
+export const NAME_VALIDATION_REGEX = /^[а-яіґїє ,.'-]+$/i
+export const PHONE_VALIDATION_REGEX =
+	/^\(?([0-9]{3})\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$/
